clean up entry point and rename
